Fixed * PIMS-903 Stock expiration report * PIMS-904 Show inventory allocated to a shipment on the stock card * PIMS-1080 Allow user to view daily transaction report * PIMS-1109 Orders should be marked as Partially Received if the order has some items have been received * PIMS-1083 Allow user to create a request * PIMS-1073 If you try to add an item with zero quantity to a shipment from the inventory browser, it silently doesn't appear * Added functionality around Requests (e.g. View Request, Add Comment, Add Document, etc).
git-svn-id: https://dev.pih-emr.org/svn/repository/projects/warehouse/trunk@10521 f0ac6d2b-dfc5-4988-9158-fc266d298d87

+++ grails-app/controllers/org/pih/warehouse/shipping/CreateShipmentWorkflowController.groovy
@@ -23,7 +23,7 @@ class CreateShipmentWorkflowController {
     			
     			// create a new shipment instance if we don't have one already
     			if (!flow.shipmentInstance) { 
-    				flow.shipmentInstance = shipmentService.getShipmentInstance(params.id as Long)
+    				flow.shipmentInstance = shipmentService.getShipmentInstance(params.id)
     				flow.shipmentWorkflow = shipmentService.getShipmentWorkflow(flow.shipmentInstance)
     			}
 				if (params.skipTo) {
